# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/conference/join-group-confernce": {
            "patch": {
                "description": "JOIN GROUP CONFERENCE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CONFERENCE"
                ],
                "summary": "API FOR JOIN GROUP CONFERENCE",
                "operationId": "JOIN-GROUP-CONFERENCE",
                "parameters": [
                    {
                        "description": "Request body for join group conference",
                        "name": "JOIN-GROUP-CONFERENCE",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.JoinGroupConferenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinGroupConferenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinGroupConferenceResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinGroupConferenceResponse"
                        }
                    }
                }
            }
        },
        "/api/conference/join-private-conference": {
            "patch": {
                "description": "JOIN PRIVATE CONFERENCE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CONFERENCE"
                ],
                "summary": "API FOR JOIN PRIVATE CONFERENCE",
                "operationId": "JOIN-PRIVATE-CONFERENCE",
                "parameters": [
                    {
                        "description": "Request body for join private conference",
                        "name": "JOIN-PRIVATE-CONFERENCE",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.JoinPrivateConferenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinPrivateConferenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinPrivateConferenceResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinPrivateConferenceResponse"
                        }
                    }
                }
            }
        },
        "/api/conference/join-public-conference": {
            "patch": {
                "description": "JOIN PUBLIC CONFERENCE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CONFERENCE"
                ],
                "summary": "API FOR JOIN PUBLIC CONFERENCE",
                "operationId": "JOIN-PUBLIC-CONFERENCE",
                "parameters": [
                    {
                        "description": "Request body for sjoin public conference",
                        "name": "JOIN-PUBLIC-CONFERENCE",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.JoinPublicConferenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinPublicConferenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinPublicConferenceResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/conference.JoinPublicConferenceResponse"
                        }
                    }
                }
            }
        },
        "/api/conference/schedule-conference": {
            "post": {
                "description": "CONFERENCE SCHEDULE CONFERENCE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CONFERENCE"
                ],
                "summary": "API FOR SCHEDULE CONFERENCE",
                "operationId": "SCHEDULE-CONFERENCE",
                "parameters": [
                    {
                        "description": "Request body for schedule conference",
                        "name": "SCHEDULE-CONFERENCE",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ScheduleConferenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conference.ScheduleConferenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conference.ScheduleConferenceResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/conference.ScheduleConferenceResponse"
                        }
                    }
                }
            }
        },
        "/api/conference/start-private-conference": {
            "post": {
                "description": "START PRIVATE CONFERENCE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CONFERENCE"
                ],
                "summary": "API FOR START PRIVATE CONFERENCE",
                "operationId": "START-PRIVATE-CONFERENCE",
                "parameters": [
                    {
                        "description": "Request body for start private conference",
                        "name": "START-PRIVATE-CONFERENCE",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.StartPrivateConferenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conference.StartPrivateConferenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conference.StartPrivateConferenceResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/conference.StartPrivateConferenceResponse"
                        }
                    }
                }
            }
        },
        "/api/conference/start-public-conference": {
            "post": {
                "description": "START PUBLIC CONFERENCE",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CONFERENCE"
                ],
                "summary": "API FOR START PUBLIC CONFERENCE",
                "operationId": "START-PUBLIC-CONFERENCE",
                "parameters": [
                    {
                        "description": "Request body for start public conference",
                        "name": "START-PUBLIC-CONFERENCE",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.StartPublicConferenceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/conference.StartPublicConferenceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/conference.StartPublicConferenceResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/conference.StartPublicConferenceResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "conference.JoinGroupConferenceResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
        "conference.JoinPrivateConferenceResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
        "conference.JoinPublicConferenceResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
...
        "conference.ScheduleConferenceResponse": {
            "type": "object",
            "properties": {
                "ConferenceID": {
                    "type": "string"
                },
                "Result": {
                    "type": "string"
                }
            }
        },
        "conference.StartPrivateConferenceResponse": {
            "type": "object",
            "properties": {
                "Result": {
                    "type": "string"
                },
                "conferenceID": {
                    "type": "string"
                }
            }
        },
        "conference.StartPublicConferenceResponse": {
            "type": "object",
            "properties": {
                "ConferenceID": {
                    "type": "string"
                },
                "Result": {
                    "type": "string"
                }
            }
        },
...
        "models.JoinGroupConferenceRequest": {
            "type": "object",
            "properties": {
                "conferenceID": {
                    "type": "string"
                },
                "groupID": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "models.JoinPrivateConferenceRequest": {
            "type": "object",
            "properties": {
                "conferenceID": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "models.JoinPublicConferenceRequest": {
            "type": "object",
            "properties": {
                "conferenceID": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "models.ScheduleConferenceRequest": {
            "type": "object",
            "properties": {
                "broadcast": {
                    "type": "boolean"
                },
                "chat": {
                    "type": "boolean"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "interest": {
                    "type": "string"
                },
                "participantlimit": {
                    "type": "integer"
                },
                "recording": {
                    "type": "boolean"
                },
                "time_nanos": {
                    "type": "integer"
                },
                "time_seconds": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "models.StartPrivateConferenceRequest": {
            "type": "object",
            "properties": {
                "broadcast": {
                    "type": "boolean"
                },
                "chat": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "interest": {
                    "type": "string"
                },
                "participantlimit": {
                    "type": "integer"
                },
                "recording": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "models.StartPublicConferenceRequest": {
            "type": "object",
            "properties": {
                "broadcast": {
                    "type": "boolean"
                },
                "chat": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "interest": {
                    "type": "string"
                },
                "joinType": {
                    "type": "string"
                },
                "participantlimit": {
                    "type": "integer"
                },
                "recording": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
    }
}